# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "description": "API documentation for Laravel + React project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/addcandidate": {
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Add a new candidate",
                "operationId": "49c9cfe63ada19ada0f1f8e71d51a03f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "candidate@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepassword"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "9876543210"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidate details added successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidate/upload-files": {
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Upload candidate files",
                "operationId": "fe9c7f6135d5f2d69710248285a28897",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "profile_photo"
                                ],
                                "properties": {
                                    "profile_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "signature": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "id_proof": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "new_me": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "other_identification": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "other_identification2": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "other_identification3": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "other_identification4": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Files uploaded successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/candidates/count": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get total number of candidates",
                "operationId": "726a5faa50241e95ec1e0893b125b341",
                "responses": {
                    "200": {
                        "description": "Total number of candidates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}